speech: wire whisper download into runtime, add streaming STT events, gut speech download module
Batch A (Units 1-4):
- runtime: add whisper_base_en() model manifest to DownloadManager
- runtime: add Step 10.5 ensure_whisper_model() before speech actors
- bus: add TranscribedWord, TranscriptionWordReady, SttModelLoaded, SttModelLoadFailed, SttListening, SttStopped, SttCancelled events
- bus: update TranscriptionCompleted with words and average_confidence fields
- speech: delete download.rs, create models.rs (path resolution only, no network)
- speech: replace run_speech_onboarding with check_speech_models
- speech: remove reqwest, sha2, hex, futures-util deps
